requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Filtering and Search
django-filter==23.3

# Image Processing
Pillow==10.1.0

# CORS Support
django-cors-headers==4.3.0

# Environment Variables
python-decouple==3.8

# Redis Cache
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-yasg==1.21.7

# Development Tools
django-debug-toolbar==4.2.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
